CLN: moved gromacs_file_tree_builder to force_gromacs.io module
