+++ force_gromacs/tests/probe_classes.py
@@ -3,7 +3,7 @@ from unittest import mock
 from force_gromacs.commands.gromacs_commands import (
     Gromacs_genbox, Gromacs_genion
 )
-from force_gromacs.commands.gromacs_file_tree_builder import (
+from force_gromacs.io.gromacs_file_tree_builder import (
     GromacsFileTreeBuilder
 )
 from force_gromacs.data_sources.fragment import Fragment
